--- backend/api/handlers.go
@@ -1,1 +1,0 @@
-package api
