--- backend/db/database.go
@@ -1,1 +1,0 @@
-package db
