+++ backend/go.mod
@@ -1,4 +1,4 @@
-module test
+module backend
 
 go 1.21.0
