+++ backend/handlers/login.go
@@ -1,0 +1,58 @@
+package handlers
+
+import (
+	"backend/models"
+	"backend/utils"
+	"github.com/gin-gonic/gin"
+	"golang.org/x/crypto/bcrypt"
+	"net/http"
+)
+
+func (u *UserController) UserLogin(ctx *gin.Context) {
+	//获取参数
+	name := ctx.PostForm("name")
+	password := ctx.PostForm("password")
+
+	//数据验证
+	if len(name) == 0 {
+		ctx.JSON(http.StatusUnprocessableEntity, gin.H{
+			"code":    422,
+			"message": "用户名不能为空",
+		})
+		return
+	}
+	if len(password) < 6 {
+		ctx.JSON(http.StatusUnprocessableEntity, gin.H{
+			"code":    422,
+			"message": "密码不能少于6位",
+		})
+		return
+	}
+
+	//判断手机号是否存在
+	var user models.User
+	u.db.Model(&models.User{}).Where("name = ?", name).First(&user)
+	if user.ID == 0 {
+		ctx.JSON(http.StatusUnprocessableEntity, gin.H{
+			"code":    422,
+			"message": "用户不存在",
+		})
+		return
+	}
+
+	//判断密码是否正确
+	if err := bcrypt.CompareHashAndPassword([]byte(user.Password), []byte(password)); err != nil {
+		ctx.JSON(http.StatusUnprocessableEntity, gin.H{
+			"code":    422,
+			"message": "密码错误",
+		})
+		return
+	}
+
+	//返回结果
+	ctx.JSON(http.StatusOK, gin.H{
+		"code":    200,
+		"message": "登录成功",
+	})
+	utils.GenerateToken(ctx, user)
+}
