+++ backend/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"backend/router"
+)
+
+func main() {
+	router.InitRouter()
+}
