--- backend/test_1.go
@@ -1,448 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"github.com/dgrijalva/jwt-go"
-	"github.com/gin-gonic/gin"
-	_ "github.com/go-sql-driver/mysql"
-	"github.com/jinzhu/gorm"
-	"golang.org/x/crypto/bcrypt"
-	"log"
-	"net/http"
-	"strings"
-	"time"
-)
-
-type User struct {
-	gorm.Model
-	Name      string `gorm:"varchar(20);not null"`
-	Telephone string `gorm:"varchar(20);not null;unique"`
-	Password  string `gorm:"size:255;not null"`
-}
-
-func main() {
-
-	//获取初始化的数据库
-	db := InitDB()
-	//延迟关闭数据库
-	defer db.Close()
-
-	//创建一个默认的路由引擎
-	r := gin.Default()
-	r.Use(JWTAuth())
-
-	//注册
-	r.POST("/register", func(ctx *gin.Context) {
-
-		//获取参数
-		name := ctx.PostForm("name")
-		telephone := ctx.PostForm("telephone")
-		password := ctx.PostForm("password")
-
-		//数据验证
-		if len(name) == 0 {
-			ctx.JSON(http.StatusUnprocessableEntity, gin.H{
-				"code":    422,
-				"message": "用户名不能为空",
-			})
-			return
-		}
-		if len(telephone) != 11 {
-			ctx.JSON(http.StatusUnprocessableEntity, gin.H{
-				"code":    422,
-				"message": "手机号必须为11位",
-			})
-			return
-		}
-		if len(password) < 6 {
-			ctx.JSON(http.StatusUnprocessableEntity, gin.H{
-				"code":    422,
-				"message": "密码不能少于6位",
-			})
-			return
-		}
-
-		//判断用户名是否存在get
-		var user User
-		db.Model(&User{}).Where("name = ?", name).First(&user)
-		if user.ID != 0 {
-			ctx.JSON(http.StatusUnprocessableEntity, gin.H{
-				"code":    422,
-				"message": "用户名已存在",
-			})
-			return
-		}
-
-		//创建用户
-		hashedPassword, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
-		if err != nil {
-			ctx.JSON(http.StatusUnprocessableEntity, gin.H{
-				"code":    500,
-				"message": "密码加密错误",
-			})
-			return
-		}
-		newUser := User{
-			Name:      name,
-			Telephone: telephone,
-			Password:  string(hashedPassword),
-		}
-		result := db.Model(&User{}).Create(&newUser)
-		if result.Error != nil {
-			ctx.JSON(http.StatusInternalServerError, gin.H{
-				"code":    500,
-				"message": "注册失败",
-			})
-			return
-		}
-
-		//返回结果
-		ctx.JSON(http.StatusOK, gin.H{
-			"code":    200,
-			"message": "注册成功",
-		})
-	})
-
-	//登录
-	r.POST("/login", func(ctx *gin.Context) {
-
-		//获取参数
-		name := ctx.PostForm("name")
-		password := ctx.PostForm("password")
-
-		//数据验证
-		if len(name) == 0 {
-			ctx.JSON(http.StatusUnprocessableEntity, gin.H{
-				"code":    422,
-				"message": "用户名不能为空",
-			})
-			return
-		}
-		if len(password) < 6 {
-			ctx.JSON(http.StatusUnprocessableEntity, gin.H{
-				"code":    422,
-				"message": "密码不能少于6位",
-			})
-			return
-		}
-
-		//判断手机号是否存在
-		var user User
-		db.Model(&User{}).Where("name = ?", name).First(&user)
-		if user.ID == 0 {
-			ctx.JSON(http.StatusUnprocessableEntity, gin.H{
-				"code":    422,
-				"message": "用户不存在",
-			})
-			return
-		}
-
-		//判断密码是否正确
-		if err := bcrypt.CompareHashAndPassword([]byte(user.Password), []byte(password)); err != nil {
-			ctx.JSON(http.StatusUnprocessableEntity, gin.H{
-				"code":    422,
-				"message": "密码错误",
-			})
-			return
-		}
-
-		//返回结果
-		ctx.JSON(http.StatusOK, gin.H{
-			"code":    200,
-			"message": "登录成功",
-		})
-		generateToken(ctx, user)
-
-	})
-
-	//修改密码
-	r.POST("/password_change", func(ctx *gin.Context) {
-
-		//获取参数
-		name := ctx.PostForm("name")
-		telephone := ctx.PostForm("telephone")
-		old_password := ctx.PostForm("old_password")
-		new_password := ctx.PostForm("new_password")
-
-		//数据验证
-		if len(name) == 0 {
-			ctx.JSON(http.StatusUnprocessableEntity, gin.H{
-				"code":    422,
-				"message": "用户名不能为空",
-			})
-			return
-		}
-		if len(telephone) != 11 {
-			ctx.JSON(http.StatusUnprocessableEntity, gin.H{
-				"code":    422,
-				"message": "手机号必须为11位",
-			})
-			return
-		}
-		if len(new_password) < 6 {
-			ctx.JSON(http.StatusUnprocessableEntity, gin.H{
-				"code":    422,
-				"message": "密码不能少于6位",
-			})
-			return
-		}
-
-		//判断用户名和手机号是否存在
-		var user User
-		db.Model(&User{}).Where("name = ? AND telephone = ?", name, telephone).First(&user)
-		if user.ID == 0 {
-			ctx.JSON(http.StatusUnprocessableEntity, gin.H{
-				"code":    422,
-				"message": "用户名或手机号错误",
-			})
-			return
-		}
-
-		//判断旧密码是否正确
-		if err := bcrypt.CompareHashAndPassword([]byte(user.Password), []byte(old_password)); err != nil {
-			ctx.JSON(http.StatusUnprocessableEntity, gin.H{
-				"code":    422,
-				"message": "旧密码错误",
-			})
-			return
-		}
-
-		//修改密码（创建用户）
-		hashednewPassword, err := bcrypt.GenerateFromPassword([]byte(new_password), bcrypt.DefaultCost)
-		if err != nil {
-			ctx.JSON(http.StatusUnprocessableEntity, gin.H{
-				"code":    500,
-				"message": "密码加密错误",
-			})
-			return
-		}
-
-		result := db.Model(&User{}).Where("name = ? AND telephone = ?", name, telephone).Update("password", string(hashednewPassword))
-		if result.Error != nil {
-			ctx.JSON(http.StatusInternalServerError, gin.H{
-				"code":    500,
-				"message": "密码更新失败",
-			})
-			return
-		}
-		if result.RowsAffected == 0 {
-			ctx.JSON(http.StatusBadRequest, gin.H{
-				"code":    400,
-				"message": "未找到匹配的用户",
-			})
-			return
-		}
-
-		//返回结果
-		ctx.JSON(http.StatusOK, gin.H{
-			"code":    200,
-			"message": "修改密码成功",
-		})
-	})
-	r.GET("/getclaim", GetClaimByTime)
-
-	//在9090端口启动服务
-	panic(r.Run(":9090"))
-}
-
-func InitDB() *gorm.DB {
-	driverName := "mysql"
-	host := "127.0.0.1"
-	port := "3306"
-	database := "users_data"
-	username := "root"
-	password := "zkq200349"
-	charset := "utf8"
-	args := fmt.Sprintf("%s:%s@(%s:%s)/%s?charset=%s&parseTime=true",
-		username,
-		password,
-		host,
-		port,
-		database,
-		charset)
-
-	db, err := gorm.Open(driverName, args)
-	if err != nil {
-		panic("failed to connect database, err:" + err.Error())
-	}
-
-	//迁移
-	db.AutoMigrate(&User{})
-
-	return db
-}
-
-// 定义一个JWTAuth的中间件
-func JWTAuth() gin.HandlerFunc {
-	return func(c *gin.Context) {
-		//过滤是否验证token， login结构直接放行，这里为了简单起见，直接判断路径中是否带login，携带login直接放行
-		if strings.Contains(c.Request.RequestURI, "login") || strings.Contains(c.Request.RequestURI, "register") {
-			return
-		}
-
-		token := c.Request.Header.Get("token")
-		if token == "" {
-			c.JSON(http.StatusOK, gin.H{
-				"status": -1,
-				"msg":    "请求未携带token，无权限访问",
-				"data":   nil,
-			})
-			c.Abort()
-			return
-		}
-
-		log.Print("get token: ", token)
-
-		// 初始化一个JWT对象实例，并根据结构体方法来解析token
-		j := NewJWT()
-		// 解析token中包含的相关信息(有效载荷)
-		claims, err := j.ParserToken(token)
-
-		if err != nil {
-			// token过期
-			if ve, ok := err.(*jwt.ValidationError); ok {
-				if ve.Errors&jwt.ValidationErrorExpired != 0 {
-					c.JSON(http.StatusOK, gin.H{
-						"status": -1,
-						"msg":    "token授权已过期，请重新申请授权",
-						"data":   nil,
-					})
-					c.Abort()
-					return
-				}
-			}
-			// 其他错误
-			c.JSON(http.StatusOK, gin.H{
-				"status": -1,
-				"msg":    err.Error(),
-				"data":   nil,
-			})
-			c.Abort()
-			return
-		}
-
-		// 将解析后的有效载荷claims重新写入gin.Context引用对象中
-		c.Set("claims", claims)
-
-	}
-}
-
-// 定义一个jwt对象
-type JWT struct {
-	// 声明签名信息
-	SigningKey []byte
-}
-
-// 初始化jwt对象
-func NewJWT() *JWT {
-	return &JWT{
-		[]byte("nearbysocial.top"),
-	}
-}
-
-// 自定义有效载荷(这里采用自定义的Name和Telephone作为有效载荷的一部分)
-type CustomClaims struct {
-	Name      string `json:"name"`
-	Telephone string `json:"telephone"`
-	// StandardClaims结构体实现了Claims接口(Valid()函数)
-	jwt.StandardClaims
-}
-
-// 调用jwt-go库生成token
-// 指定编码的算法为jwt.SigningMethodHS256
-func (j *JWT) CreateToken(claims CustomClaims) (string, error) {
-	// https://gowalker.org/github.com/dgrijalva/jwt-go#Token
-	// 返回一个token的结构体指针
-	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
-	return token.SignedString(j.SigningKey)
-}
-
-// token解码
-func (j *JWT) ParserToken(tokenString string) (*CustomClaims, error) {
-	// https://gowalker.org/github.com/dgrijalva/jwt-go#ParseWithClaims
-	// 输入用户自定义的Claims结构体对象,token,以及自定义函数来解析token字符串为jwt的Token结构体指针
-	// Keyfunc是匿名函数类型: type Keyfunc func(*Token) (interface{}, error)
-	// func ParseWithClaims(tokenString string, claims Claims, keyFunc Keyfunc) (*Token, error) {}
-	token, err := jwt.ParseWithClaims(tokenString, &CustomClaims{}, func(token *jwt.Token) (interface{}, error) {
-		return j.SigningKey, nil
-	})
-
-	if err != nil {
-		// https://gowalker.org/github.com/dgrijalva/jwt-go#ValidationError
-		// jwt.ValidationError 是一个无效token的错误结构
-		if ve, ok := err.(*jwt.ValidationError); ok {
-			// ValidationErrorMalformed是一个uint常量，表示token不可用
-			if ve.Errors&jwt.ValidationErrorMalformed != 0 {
-				return nil, fmt.Errorf("token不可用")
-				// ValidationErrorExpired表示Token过期
-			} else if ve.Errors&jwt.ValidationErrorExpired != 0 {
-				return nil, fmt.Errorf("token过期")
-				// ValidationErrorNotValidYet表示无效token
-			} else if ve.Errors&jwt.ValidationErrorNotValidYet != 0 {
-				return nil, fmt.Errorf("无效的token")
-			} else {
-				return nil, fmt.Errorf("token不可用")
-			}
-
-		}
-	}
-
-	// 将token中的claims信息解析出来并断言成用户自定义的有效载荷结构
-	if claims, ok := token.Claims.(*CustomClaims); ok && token.Valid {
-		return claims, nil
-	}
-
-	return nil, fmt.Errorf("token无效")
-
-}
-
-// token生成器
-func generateToken(c *gin.Context, user User) {
-	// 构造SignKey: 签名和解签名需要使用一个值
-	j := NewJWT()
-
-	// 构造用户claims信息(负荷)
-	claims := CustomClaims{
-		user.Name,
-		user.Telephone,
-		jwt.StandardClaims{
-			NotBefore: int64(time.Now().Unix() - 1000), // 签名生效时间
-			ExpiresAt: int64(time.Now().Unix() + 3600), // 签名过期时间
-			Issuer:    "NearbySocial",                  // 签名颁发者
-		},
-	}
-
-	// 根据claims生成token对象
-	token, err := j.CreateToken(claims)
-
-	if err != nil {
-		c.JSON(http.StatusOK, gin.H{
-			"status":  -1,
-			"message": err.Error(),
-			"data":    nil,
-		})
-	}
-
-	log.Println(token)
-	c.JSON(http.StatusOK, gin.H{
-		"token": token,
-	})
-
-	return
-
-}
-
-// 定义一个普通controller函数，作为一个验证接口逻辑
-func GetClaimByTime(c *gin.Context) {
-	// 上面我们在JWTAuth()中间中将'claims'写入到gin.Context的指针对象中，因此在这里可以将之解析出来
-	claims := c.MustGet("claims").(*CustomClaims)
-	if claims != nil {
-		c.JSON(http.StatusOK, gin.H{
-			"status":  0,
-			"message": "token有效",
-			"data":    claims,
-		})
-	}
-}
-
-//just test
