+++ backend/utils/jwt.go
@@ -1,0 +1,116 @@
+package utils
+
+import (
+	"backend/models"
+	"fmt"
+	"github.com/dgrijalva/jwt-go"
+	"github.com/gin-gonic/gin"
+	"log"
+	"net/http"
+	"time"
+)
+
+// 定义一个jwt对象
+type JWT struct {
+	// 声明签名信息
+	SigningKey []byte
+}
+
+// 初始化jwt对象
+func NewJWT() *JWT {
+	return &JWT{
+		[]byte("nearbysocial.top"),
+	}
+}
+
+// 自定义有效载荷(这里采用自定义的Name和Telephone作为有效载荷的一部分)
+type CustomClaims struct {
+	Name      string `json:"name"`
+	Telephone string `json:"telephone"`
+	// StandardClaims结构体实现了Claims接口(Valid()函数)
+	jwt.StandardClaims
+}
+
+// 调用jwt-go库生成token
+// 指定编码的算法为jwt.SigningMethodHS256
+func (j *JWT) CreateToken(claims CustomClaims) (string, error) {
+	// https://gowalker.org/github.com/dgrijalva/jwt-go#Token
+	// 返回一个token的结构体指针
+	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
+	return token.SignedString(j.SigningKey)
+}
+
+// token解码
+func (j *JWT) ParserToken(tokenString string) (*CustomClaims, error) {
+	// https://gowalker.org/github.com/dgrijalva/jwt-go#ParseWithClaims
+	// 输入用户自定义的Claims结构体对象,token,以及自定义函数来解析token字符串为jwt的Token结构体指针
+	// Keyfunc是匿名函数类型: type Keyfunc func(*Token) (interface{}, error)
+	// func ParseWithClaims(tokenString string, claims Claims, keyFunc Keyfunc) (*Token, error) {}
+	token, err := jwt.ParseWithClaims(tokenString, &CustomClaims{}, func(token *jwt.Token) (interface{}, error) {
+		return j.SigningKey, nil
+	})
+
+	if err != nil {
+		// https://gowalker.org/github.com/dgrijalva/jwt-go#ValidationError
+		// jwt.ValidationError 是一个无效token的错误结构
+		if ve, ok := err.(*jwt.ValidationError); ok {
+			// ValidationErrorMalformed是一个uint常量，表示token不可用
+			if ve.Errors&jwt.ValidationErrorMalformed != 0 {
+				return nil, fmt.Errorf("token不可用")
+				// ValidationErrorExpired表示Token过期
+			} else if ve.Errors&jwt.ValidationErrorExpired != 0 {
+				return nil, fmt.Errorf("token过期")
+				// ValidationErrorNotValidYet表示无效token
+			} else if ve.Errors&jwt.ValidationErrorNotValidYet != 0 {
+				return nil, fmt.Errorf("无效的token")
+			} else {
+				return nil, fmt.Errorf("token不可用")
+			}
+
+		}
+	}
+
+	// 将token中的claims信息解析出来并断言成用户自定义的有效载荷结构
+	if claims, ok := token.Claims.(*CustomClaims); ok && token.Valid {
+		return claims, nil
+	}
+
+	return nil, fmt.Errorf("token无效")
+
+}
+
+// token生成器
+func GenerateToken(c *gin.Context, user models.User) {
+	// 构造SignKey: 签名和解签名需要使用一个值
+	j := NewJWT()
+
+	// 构造用户claims信息(负荷)
+	claims := CustomClaims{
+		user.Name,
+		user.Telephone,
+		jwt.StandardClaims{
+			NotBefore: int64(time.Now().Unix() - 1000), // 签名生效时间
+			ExpiresAt: int64(time.Now().Unix() + 3600), // 签名过期时间
+			Issuer:    "NearbySocial",                  // 签名颁发者
+		},
+	}
+
+	// 根据claims生成token对象
+	token, err := j.CreateToken(claims)
+
+	if err != nil {
+		c.JSON(http.StatusOK, gin.H{
+			"status":  -1,
+			"message": err.Error(),
+			"data":    nil,
+		})
+	}
+
+	log.Println(token)
+	c.JSON(http.StatusOK, gin.H{
+		"token": token,
+	})
+
+	return
+
+}
