--- backend/websocket/hub.go
@@ -1,1 +1,0 @@
-package websocket
